pyproject.toml
[project]
name = "mcp-testing"
version = "0.1.0"
description = "AI Assistant POC with MCP tools, SQLite chat history, and React UI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastmcp>=3.1.0",
    "google-genai>=1.66.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "ddgs>=9.11.2",
]
